+++ Frontend/src/App.jsx
@@ -5,8 +5,7 @@ import React, {
   useState,
   Navigate,
 } from "react";
-import Cookies from "js-cookie";
-import jwt_decode from "jwt-decode";
+
 import ProfileBar from "./components/NavAuthDemo";
 import LoginPage from "./pages/auth/LoginPage";
 
@@ -15,44 +14,20 @@ import "./App.css";
 import MainPage from "./pages/main/MainPageDemo";
 import RegisterPage from "./pages/auth/RegisterPage";
 import AssignPage from "./pages/auth/AssignPage";
+import checkToken from "./services/checkToken";
+import Cookies from "js-cookie";
 
 const UserStateContext = createContext();
 
 function App() {
+  const authToken = Cookies.get("authToken");
+  console.log("Cookie_token :  ", authToken);
   const [userState, setUserState] = useState(null); // login?
   const [enableAssignPage, setEnableAssignPage] = useState(false); // login?
   console.log(userState);
+  console.log(Cookies.fet);
   useEffect(() => {
-    try {
-      const authToken = Cookies.get("authToken");
-      // console.log("authToken", authToken);
-      if (!authToken) {
-        console.log("Cookie_token not found");
-        const g_token = localStorage.getItem("accessToken");
-        if (g_token) {
-          localStorage.removeItem("accessToken");
-        }
-      } else {
-        // use api login user by sent mail and google token for verify
-        const authToken = Cookies.get("authToken");
-        const decodedToken = jwt_decode(authToken);
-
-        if (decodedToken) {
-          // console.log("DecodedToken : ", decodedToken);
-          setUserState(decodedToken);
-
-          // window.location.reload();
-        } else {
-          // console.log("DecodedToken : ", decodedToken);
-          console.log("Token expired or Token not found or Token invalid");
-          Cookies.remove("authToken");
-          // window.location.reload();
-        }
-      }
-    } catch (e) {
-      console.log(e);
-      Cookies.remove("authToken");
-    }
+    checkToken(setUserState);
   }, []);
   return (
     // userState is data of user from token that decoded
